moved Point class to seperate file
