+++ src/main/java/org/avphs/racingline/Point.java
@@ -1,0 +1,40 @@
+package org.avphs.racingline;
+
+class Point {
+    int x, y;
+
+    public Point() {
+        x = 0;
+        y = 0;
+    }
+
+    public Point(int _x, int _y) {
+        x = _x;
+        y = _y;
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + x;
+        result = prime * result + y;
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        Point other = (Point) obj;
+        if (x != other.x)
+            return false;
+        if (y != other.y)
+            return false;
+        return true;
+    }
+}
